requirements.txt
numpy==1.19.5
tensorflow==2.8.0
opencv-python==4.5.3.56
scikit-learn==0.24.2
matplotlib
Pillow
fastapi
uvicorn
pandas